Realice el put del products y organice vistas
