+++ src/front/js/pages/DashboardViews/dashMenuView.jsx
@@ -1,0 +1,24 @@
+import React, { useContext, useEffect } from "react";
+import { Context } from "../../store/appContext";
+import { Navigate } from "react-router-dom";
+import DashMenu from "../../component/dashboard/dashmenu.jsx";
+
+const MyDashMenuView = () => {
+  const { store, actions } = useContext(Context);
+
+  useEffect(() => {
+    actions.getProductsDetails(localStorage.getItem("costumer_id"));
+  }, [store.products]);
+
+  return localStorage.getItem("costumer_id") === null ? (
+    <Navigate to="/login" />
+  ) : (
+    <div className="container-fluid row g-4 my-2 d-flex justify-content-center">
+      <div className="col-12">
+        <DashMenu />
+      </div>
+    </div>
+  );
+};
+
+export default MyDashMenuView;
